Makes input boxes of search plugin work better on mobile
This change makes the input boxes of the search plugin work better on mobile. Changing the input type to "search" makes it so that the virtual keyboard that pops up on mobile devices has a search button where the enter key normally is, so that clicking in will execute the search. Currently, with the type set to "text", the enter button on mobile keyboards is changed into a tab, because the browser sees, the input box for the search in the wiki footer, and thinks the user might want to tab over there instead of executing the search.